add sendText node
