+++ .github/gulp-templates/new-node/TNameNodeConfig.ts.template
@@ -1,3 +1,3 @@
-import { TNodeConfig } from './TNodeConfig';
+import { TVBANNodeConfig } from './TVBANNodeConfig';
 
-export type T@@CAMEL_NODE_NAMENodeConfig = TNodeConfig;
+export type T@@CAMEL_NODE_NAMENodeConfig = TVBANNodeConfig;
