+++ .github/gulp-templates/new-node/name.ts.template
@@ -1,4 +1,4 @@
-import { Node } from '../lib/Node';
+import { VBANNode } from '../lib/VBANNode';
 import * as REDRegistry from '@node-red/registry';
 import { registerNode } from '../lib/registerNode';
 import { T@@CAMEL_NODE_NAMENode } from '../types/T@@CAMEL_NODE_NAMENode';
@@ -6,7 +6,7 @@ import { T@@CAMEL_NODE_NAMENodeConfig } from '../types/T@@CAMEL_NODE_NAMENodeCon
 
 const NODE_NAME = '@@KEBAB_NODE_NAME';
 
-class @@CAMEL_NODE_NAME extends Node<T@@CAMEL_NODE_NAMENode, T@@CAMEL_NODE_NAMENodeConfig> {
+class @@CAMEL_NODE_NAME extends VBANNode<T@@CAMEL_NODE_NAMENode, T@@CAMEL_NODE_NAMENodeConfig> {
     protected async init(): Promise<void> {
         await super.init();
     }
