+++ src/nodes/VBANSender.ts
@@ -3,48 +3,20 @@ import { registerNode } from '../lib/registerNode';
 import { VBANNode } from '../lib/VBANNode';
 import { ENodeStatus } from '../lib/ENodeStatus';
 import { NodeMessageInFlow } from 'node-red';
-import {
-    EFormatBit,
-    ESerialStreamType,
-    ESubProtocol,
-    ETextEncoding,
-    ISerialBitMode,
-    VBANPacket,
-    VBANSerialPacket,
-    VBANTEXTPacket
-} from 'vban';
 import { TVBANSenderNode } from '../types/TVBANSenderNode';
 import { TVBANSenderNodeConfig } from '../types/TVBANSenderNodeConfig';
-
-export type TSubProtocol = 'AUDIO' | 0 | 'SERIAL' | 1 | 'TXT' | 2 | 'SERVICE' | 3 | ESubProtocol;
-
-export type TStreamTypeStr = 'GENERIC' | 'MIDI';
-
-export interface IPacket {
-    subProtocol: TSubProtocol;
-    streamName: string;
-    sr?: number;
-    frameCounter?: number;
-    data?: Buffer;
-    //serial specifics
-    bitMode?: ISerialBitMode;
-    channelsIdents?: number;
-    bps?: number;
-    formatBit?: EFormatBit;
-    streamType?: TStreamTypeStr | ESerialStreamType;
-    //text specifics
-    text?: string;
-    encoding?: ETextEncoding;
-    //TODO add AUDIO and SERVICE
-}
+import { IPacket, Sender } from '../lib/Sender';
 
 const NODE_NAME = 'vban-sender';
 
 class VBANSender extends VBANNode<TVBANSenderNode, TVBANSenderNodeConfig> {
+    private sender?: Sender;
     protected async init(): Promise<void> {
         await super.init();
 
         if (this.serverConfigured) {
+            this.sender = new Sender(this.server);
+
             this.node.on('input', (msg) => {
                 const { payload } = msg as NodeMessageInFlow & {
                     payload?: {
@@ -60,80 +32,6 @@ class VBANSender extends VBANNode<TVBANSenderNode, TVBANSenderNodeConfig> {
         } else {
             this.setStatus(ENodeStatus.ERROR, 'VBAN server not configured');
         }
-    }
-
-    private getSerialStreamType(streamType?: ESerialStreamType | TStreamTypeStr): ESerialStreamType {
-        if (!streamType && streamType !== 0) {
-            throw new Error('streamType is needed');
-        }
-
-        if (Number(streamType) && ESerialStreamType[Number(streamType)]) {
-            return streamType as ESerialStreamType;
-        } else {
-            switch (streamType) {
-                case 'GENERIC':
-                case ESerialStreamType.VBAN_SERIAL_GENERIC:
-                    return ESerialStreamType.VBAN_SERIAL_GENERIC;
-                case 'MIDI':
-                case ESerialStreamType.VBAN_SERIAL_MIDI:
-                    return ESerialStreamType.VBAN_SERIAL_MIDI;
-                default:
-                    throw new Error(`unknown stream type ${streamType}`);
-            }
-        }
-    }
-
-    private constructVBANObject(packet: Partial<IPacket> & { streamName: string; subProtocol: TSubProtocol }): VBANPacket {
-        let returnPacket: VBANPacket;
-        switch (typeof packet.subProtocol === 'string' ? packet.subProtocol?.toUpperCase() : packet.subProtocol) {
-            case 'AUDIO':
-            case 0:
-            case ESubProtocol.AUDIO:
-                throw new Error(`"${packet.subProtocol}" NOT YET IMPLEMENTED`);
-                break;
-            case 'SERIAL':
-            case 1:
-            case ESubProtocol.SERIAL:
-                if (!packet.data) {
-                    throw new Error('data are mandatory');
-                }
-
-                returnPacket = new VBANSerialPacket(
-                    {
-                        streamName: packet.streamName,
-                        formatBit: EFormatBit.VBAN_DATATYPE_BYTE8,
-                        streamType: this.getSerialStreamType(packet.streamType),
-                        bps: packet.bps as number,
-                        bitMode: packet.bitMode as ISerialBitMode,
-                        sr: packet.sr as number,
-                        channelsIdents: packet.channelsIdents as number
-                    },
-                    packet.data
-                );
-
-                break;
-            case 'TXT':
-            case 2:
-            case ESubProtocol.TEXT:
-                returnPacket = new VBANTEXTPacket(
-                    {
-                        formatBit: EFormatBit.VBAN_DATATYPE_BYTE8,
-                        streamName: packet.streamName,
-                        streamType: packet.encoding ?? ETextEncoding.VBAN_TXT_UTF8
-                    },
-                    packet.text
-                );
-                break;
-            case 'SERVICE':
-            case 3:
-            case ESubProtocol.SERVICE:
-                throw new Error(`"${packet.subProtocol}" NOT YET IMPLEMENTED`);
-                break;
-            default:
-                throw new Error(`unknown subProtocol "${packet.subProtocol}"`);
-        }
-
-        return returnPacket;
     }
 
     private sendVBANPacket(payload?: {
@@ -184,8 +82,10 @@ class VBANSender extends VBANNode<TVBANSenderNode, TVBANSenderNodeConfig> {
         }
 
         try {
-            const newPacket = this.constructVBANObject(packet as IPacket);
-            this.server.send(newPacket, Number(destination.port as number), destination.address);
+            this.sender?.send(packet as IPacket, {
+                address: destination.address,
+                port: Number(destination.port)
+            });
         } catch (e) {
             if (e instanceof Error) {
                 this.node.error(e.message, {
