+++ src/types/TSendTextMsg.ts
@@ -1,0 +1,13 @@
+import * as RED from 'node-red';
+
+export type TSendTextMsg = RED.NodeMessageInFlow &
+    Partial<{
+        payload:
+            | string
+            | number
+            | {
+                  packet?: {
+                      text: string;
+                  };
+              };
+    }>;
