+++ src/types/TSendTextNode.ts
@@ -1,0 +1,3 @@
+import * as RED from 'node-red';
+
+export type TSendTextNode = RED.Node;
