+++ src/types/TSendTextNodeConfig.ts
@@ -1,0 +1,7 @@
+import { TVBANNodeConfig } from './TVBANNodeConfig';
+
+export type TSendTextNodeConfig = TVBANNodeConfig & {
+    streamName: string;
+    address: string;
+    port: string | number;
+};
